refine rendering of place, od, user analysis
bug fix in rendering only selected analyses, change conf. interval to display of margin_or_error

+++ dp_mobility_report/model/place_analysis.py
@@ -69,7 +69,7 @@ def get_visits_per_tile(
                 visits_per_tile["visits"] / visists_sum * record_count
             ).astype(int)
             n_outliers = int(n_outliers / visists_sum * record_count)
-            moe = int(moe / visists_sum * record_count)
+            moe = moe / visists_sum * record_count
 
     # as counts are already dp, no further privacy mechanism needed
     dp_quartiles = visits_per_tile.visits.describe()
